requirements.txt
appdirs==1.4.3
Django==1.9
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
Markdown==2.6.8
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.1.10
six==1.10.0
South==1.0.2
pillow==5.2.0